adding control of LRU_GEN_STALE_SWP_ENTRY_SAVIOR

+++ mm/workingset.c
@@ -204,7 +204,7 @@ bool entry_is_entry_ext(const void *entry){
 			return false;
 		}
 		if (!xa_is_value(((struct shadow_entry*)entry)->shadow)){
-			pr_err("entry_is_entry_ext !xa_is_value ext[%lx]",entry);
+			pr_err("entry_is_entry_ext !xa_is_value ext[%lx]",(unsigned long)entry);
 			return false;
 		}
 		return true;
@@ -257,11 +257,11 @@ static void *pack_shadow_ext(int memcgid, pg_data_t *pgdat, unsigned long evicti
 #endif
 	}
 	else{
-		pr_err("entry_ext[%lx]->shadow[%lx]", entry_ext, eviction);
+		pr_err("entry_ext[%lx]->shadow[%lx]", (unsigned long)entry_ext, eviction);
 		return NULL;
 	}
 	if (((unsigned long)entry_ext & 3) != 0) {
-		pr_err("xarray err entry_ext[%lx]->shadow[%lx]", entry_ext, eviction);
+		pr_err("xarray err entry_ext[%lx]->shadow[%lx]", (unsigned long)entry_ext, eviction);
 	}
 	return entry_ext;////xa_mk_value((unsigned long)entry_ext);
 }
